// types/info.ts
export interface InfoProps {
    name: string
    metaTitle: string
    metaDescription: string
    metaImage: string
    introSafe: string
    logo: string
    siteUrl: string
    email: string
    phone: string
    address: string
    facebook: string
    twitter: string
    youtube: string
    github: string
}